fix(codegen): INFRA-316 move generateCondoSchema to codegen package
