+++ packages/codegen/package.json
@@ -5,6 +5,8 @@
     "repository": "https://github.com/open-condo-software/condo/tree/main/packages/codegen",
     "dependencies": {
         "@apollo/client": "3.3.20",
+        "@graphql-tools/graphql-file-loader": "^7.5.5",
+        "@graphql-tools/load": "^8.0.0",
         "@graphql-typed-document-node/core": "^3.1.1",
         "@open-condo/config": "workspace:^",
         "@open-condo/keystone": "workspace:^",
